chore: move validation package into internal and some minor tidy
